+++ examples/basic/ctr_en.py
@@ -5,7 +5,7 @@ from amaranth.back import verilog
 
 class Counter(Elaboratable):
     def __init__(self, width):
-        self.v = Signal(width, reset=2**width-1)
+        self.v = Signal(width, init=2**width-1)
         self.o = Signal()
         self.en = Signal()