+++ tests/test_lib_cdc.py
@@ -35,10 +35,10 @@ class FFSynchronizerTestCase(FHDLTestCase):
         sim.add_process(process)
         sim.run()
 
-    def test_reset_value(self):
-        i = Signal(reset=1)
+    def test_init_value(self):
+        i = Signal(init=1)
         o = Signal()
-        frag = FFSynchronizer(i, o, reset=1)
+        frag = FFSynchronizer(i, o, init=1)
 
         sim = Simulator(frag)
         sim.add_clock(1e-6)
@@ -53,6 +53,34 @@ class FFSynchronizerTestCase(FHDLTestCase):
             self.assertEqual((yield o), 0)
         sim.add_process(process)
         sim.run()
+
+    def test_reset_value(self):
+        i = Signal(init=1)
+        o = Signal()
+        with self.assertWarnsRegex(DeprecationWarning,
+                r"^`reset=` is deprecated, use `init=` instead$"):
+            frag = FFSynchronizer(i, o, reset=1)
+
+        sim = Simulator(frag)
+        sim.add_clock(1e-6)
+        def process():
+            self.assertEqual((yield o), 1)
+            yield i.eq(0)
+            yield Tick()
+            self.assertEqual((yield o), 1)
+            yield Tick()
+            self.assertEqual((yield o), 1)
+            yield Tick()
+            self.assertEqual((yield o), 0)
+        sim.add_process(process)
+        sim.run()
+
+    def test_reset_wrong(self):
+        i = Signal(init=1)
+        o = Signal()
+        with self.assertRaisesRegex(ValueError,
+                r"^Cannot specify both `reset` and `init`$"):
+            FFSynchronizer(i, o, reset=1, init=1)
 
 
 class AsyncFFSynchronizerTestCase(FHDLTestCase):
@@ -115,7 +143,7 @@ class AsyncFFSynchronizerTestCase(FHDLTestCase):
             sim.run()
 
     def test_neg_edge(self):
-        i = Signal(reset=1)
+        i = Signal(init=1)
         o = Signal()
         m = Module()
         m.domains += ClockDomain("sync")
@@ -166,7 +194,7 @@ class ResetSynchronizerTestCase(FHDLTestCase):
         m = Module()
         m.domains += ClockDomain("sync")
         m.submodules += ResetSynchronizer(arst)
-        s = Signal(reset=1)
+        s = Signal(init=1)
         m.d.sync += s.eq(0)
 
         sim = Simulator(m)